feat(phase-e2): MC 1.3-1.4 bracket — wither boss, witch/wither-skeleton/bat, beacon, ender chest, adventure mode, lava fluid, effects system
Version-evolution Phase 2 (Adventure Features). All values live-verified
2026-09-06 against minecraft.wiki (research record:
docs/research/phase2-1.3-1.4-research.md, ~120 VERIFIED citations).

- wither.rs: summon (T of soul sand + 3 skulls), 220-tick charge, 300 HP,
  1 HP/20-t regen, 2 s skull cadence (8 HP + Wither II 10 s Normal / 40 s
  Hard), 3x4x3 block breaking on damage, nether star + 50 XP drops
- effects.rs: timed Wither/Poison/Regeneration + beacon stat effects
  (speed/strength/resistance/jump multipliers per the w/Effect rows)
- beacon.rs: pyramid scan 1-4 (9/34/83/164 blocks), power gating,
  4 s reapply at 9+2xlevel s, range 20/30/40/50, feed cycle
- mobs: wither skeleton (fortress platform #2 spawner), witch (~0.97%
  monster pool), bat (ambient, light<=3 below sea level, groups of 8)
- ender chest: shared 27-slot container, breaks into 8 obsidian
- adventure mode (GameType 2): no block break/place, interactions open
- anvil: 3 damage stages (12%/use), gravity block; repair GUI deferred
  (no damageable items — disclosed)
- lava fluid: dimension-aware spread (OW 3/30-tick, Nether 7/10-tick),
  fluid-quad meshing with the fixed lava tint, 4 HP/10-tick damage
- emerald ore: Mountains-only generation, drops 1 emerald + 3-7 XP
- foods: potato/baked/carrot/pumpkin pie (verified hunger values)
- u16 state-space widening (the <=255 window was exhausted) + latent E1
  World::set_block state-storage bug fixed
- render-distance slider 2-32 (VERIFICATION-REPORT fix)

Suite 342 -> 372 (+30), wasm32 clean.

+++ voxelcraft/crates/vc-chunk/src/chunk.rs
@@ -356,9 +356,18 @@ impl Chunk {
     }
 
     #[inline]
-    pub fn get(&self, x: usize, y: usize, z: usize) -> u8 {
+    /// Raw BLOCK-STATE id at a position. Phase E2: widened u8 → u16 —
+    /// the E1 bracket exhausted the ≤255 state window (0..=255 fully
+    /// allocated), and E2 world blocks (anvil damage states, ender chest,
+    /// beacon, wall, frame, tripwire) need states ≥ 256. The u16 is the
+    /// raw state; callers wanting a BLOCK id fold via
+    /// `vc_blocks::blocks::state_block`. All previous call sites either
+    /// already folded (`state_block(c.get(...) as u16)` — cast now a
+    /// no-op) or compared against identity-mapped ids ≤ 56, which equals
+    /// the raw state; both keep working unchanged.
+    pub fn get(&self, x: usize, y: usize, z: usize) -> u16 {
         match &self.sections[y >> 4] {
-            Some(s) => s.get(x, y & 15, z) as u8,
+            Some(s) => s.get(x, y & 15, z),
             None => 0,
         }
     }
@@ -395,8 +404,9 @@ impl Chunk {
     }
 
     /// flat-index helpers for queued/pending edits (idx = (y<<8)|(z<<4)|x)
+    /// — raw u16 states (Phase E2 widening, see `get`)
     #[inline]
-    pub fn get_idx(&self, i: usize) -> u8 {
+    pub fn get_idx(&self, i: usize) -> u16 {
         self.get(i & 15, (i >> 8) & 0xFF, (i >> 4) & 15)
     }
